fetch all events banners in admin dashboard

+++ conference/src/main/java/com/xworkz/conference/dao/conferenceHosterDAO/ConferenceHosterDAO.java
@@ -16,5 +16,5 @@ public interface ConferenceHosterDAO {
 
     List<DelegatesEmailEntity> getAllDelegates();
 
-    ConferenceHosterEntity getAllConferenceHosterById(int conferenceId);
+    ConferenceHosterEntity getAllConferenceHosterById(Long conferenceId);
 }
